xkcd-explainer/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.3; python_version >= '3.8'
aiohttp-client-cache==0.11.0; python_version >= '3.8' and python_version < '4.0'
aiosignal==1.3.1; python_version >= '3.7'
aiosqlite==0.20.0; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
frozenlist==1.4.1; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
multidict==6.0.5; python_version >= '3.7'
pytelegrambotapi==4.16.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
soupsieve==2.5; python_version >= '3.8'
typing-extensions==4.10.0; python_version >= '3.8'
url-normalize==1.4.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
urllib3==2.2.1; python_version >= '3.8'
yarl==1.9.4; python_version >= '3.7'
